# openapi.yaml
openapi: 3.0.3
info:
  title: DataShare API
  version: 1.0.0
  description: >
    Contrat d'interface de DataShare : partage de fichiers par lien temporaire.
    Le frontend (React) communique avec le backend (NestJS) via cette API REST en JSON.
    Les routes metier sont versionnees sous /v1. L'authentification se fait par JWT
    depose dans un cookie HttpOnly (avec repli sur l'en-tete Authorization: Bearer pour
    les clients API). Les routes de telechargement public (/d/...) et le /health ne sont
    pas versionnees et ne demandent pas de compte : c'est la connaissance du lien (un UUID
    non devinable) qui fait l'autorisation.

servers:
  - url: http://localhost:3000
    description: Developpement local

tags:
  - name: Auth
    description: Inscription et connexion
  - name: Fichiers
    description: Depot, liste et suppression (authentifie)
  - name: Telechargement public
    description: Acces a un fichier via son lien (sans compte)
  - name: Systeme
    description: Sante du service

paths:
  /health:
    get:
      tags: [Systeme]
      summary: Etat de sante de l'API (health check)
      description: Utilise par Docker pour verifier que l'API repond. N'accede pas a la base.
      responses:
        '200':
          description: Le service repond
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
              example:
                status: ok
                service: datashare-api

  /v1/auth/register:
    post:
      tags: [Auth]
      summary: Creer un compte
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
            example:
              email: noemie@example.com
              password: motdepasse123
      responses:
        '201':
          description: Compte cree, jeton JWT renvoye
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Email invalide ou mot de passe < 8 caracteres
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email deja utilise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Se connecter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
            example:
              email: noemie@example.com
              password: motdepasse123
      responses:
        '201':
          description: Connexion reussie, jeton JWT renvoye
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Identifiants invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/auth/logout:
    post:
      tags: [Auth]
      summary: Se deconnecter
      description: Supprime le cookie HttpOnly d'authentification cote serveur.
      responses:
        '200':
          description: Deconnecte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  /v1/files/upload:
    post:
      tags: [Fichiers]
      summary: Televerser un fichier
      description: >
        Depose un fichier (max 1 Go), genere un lien de telechargement unique.
        Le mot de passe et la duree d'expiration sont optionnels.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Le fichier a televerser (max 1 Go)
                password:
                  type: string
                  minLength: 6
                  description: Mot de passe optionnel pour proteger le fichier
                expires_in_days:
                  type: integer
                  minimum: 1
                  maximum: 7
                  default: 7
                  description: Duree de validite du lien, entre 1 et 7 jours
      responses:
        '201':
          description: Fichier televerse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Fichier trop volumineux, executable detecte, mot de passe trop court ou duree invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Jeton JWT absent ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/files:
    get:
      tags: [Fichiers]
      summary: Lister mes fichiers
      description: Renvoie les fichiers de l'utilisateur connecte, filtrables par etat.
      security:
        - bearerAuth: []
      parameters:
        - name: filter
          in: query
          required: false
          schema:
            type: string
            enum: [active, expired]
          description: "active = liens encore valides ; expired = liens expires ; absent = tous"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
          description: "Numero de page (pagination)"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
          description: "Nombre d'elements par page"
      responses:
        '200':
          description: Liste paginée des fichiers
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/FileListItem'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
                  totalPages:
                    type: integer
        '401':
          description: Jeton JWT absent ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/files/{id}:
    delete:
      tags: [Fichiers]
      summary: Supprimer un de mes fichiers
      description: Supprime le fichier (disque + base). Seul le proprietaire peut le faire.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Fichier supprime
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: Fichier supprime
        '401':
          description: Jeton JWT absent ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Fichier introuvable (ou n'appartient pas a l'utilisateur)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /d/{token}:
    get:
      tags: [Telechargement public]
      summary: Consulter les infos d'un lien de partage
      description: Renvoie les metadonnees du fichier associe au lien. Aucun compte requis.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Le download_token (UUID) recu a l'upload
      responses:
        '200':
          description: Metadonnees du fichier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileInfo'
        '403':
          description: Ce lien a expire
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Lien invalide (aucun fichier ne correspond)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /d/{token}/download:
    post:
      tags: [Telechargement public]
      summary: Telecharger le fichier
      description: >
        Renvoie le contenu binaire du fichier en streaming. Si le fichier est protege,
        le mot de passe doit etre fourni dans le corps de la requete.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadRequest'
            example:
              password: secret1
      responses:
        '200':
          description: "Flux binaire du fichier (en-tete Content-Disposition: attachment)"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '403':
          description: Lien expire ou mot de passe incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Lien invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Jeton recu a la connexion, a envoyer dans l'en-tete : Authorization: Bearer <jeton>"

  schemas:
    Credentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email

    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        access_token:
          type: string
          description: Jeton JWT (valable 7 jours)
      example:
        user:
          id: 884308b6-d13f-4e17-9479-5a7484e9cb69
          email: noemie@example.com
        access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...

    UploadResponse:
      type: object
      properties:
        file_id:
          type: string
          format: uuid
        download_token:
          type: string
          format: uuid
        download_url:
          type: string
          format: uri
        expires_at:
          type: string
          format: date-time
      example:
        file_id: 3d79dd78-b9b5-2cd5-63ee-648a29c66ac7
        download_token: c9c3af62-31a0-3b73-9f02-db1dc3557940
        download_url: http://localhost:3000/d/c9c3af62-31a0-3b73-9f02-db1dc3557940
        expires_at: 2026-07-04T22:00:00.000Z

    FileListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        originalName:
          type: string
        sizeBytes:
          type: integer
        uploadedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        downloadToken:
          type: string
          format: uuid
        hasPassword:
          type: boolean

    FileInfo:
      type: object
      properties:
        file_name:
          type: string
        size:
          type: integer
        mime_type:
          type: string
        expires_at:
          type: string
          format: date-time
        has_password:
          type: boolean
      example:
        file_name: rapport.pdf
        size: 482190
        mime_type: application/pdf
        expires_at: 2026-07-04T22:00:00.000Z
        has_password: true

    DownloadRequest:
      type: object
      properties:
        password:
          type: string
          description: Requis seulement si le fichier est protege

    Message:
      type: object
      properties:
        message:
          type: string

    Health:
      type: object
      properties:
        status:
          type: string
        service:
          type: string

    Error:
      type: object
      description: Format d'erreur standard de NestJS
      properties:
        statusCode:
          type: integer
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        error:
          type: string
      example:
        statusCode: 403
        message: Ce lien a expire
        error: Forbidden